Added convert functions to Resource Fixed bug in types for binary string casting
