+++ test/schema.js
@@ -225,170 +225,6 @@ describe('Models', () => {
     })
   })
 
-  it('convert row', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const convertedRow = schema.convertRow('string', '10.0', '1', 'string',
-                                             'string')
-      assert.deepEqual(['string', '10.0', 1, 'string', 'string'], convertedRow)
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with less items than headers count', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      assert.throws(() => {
-        schema.convertRow('string', '10.0', '1', 'string')
-      }, Error)
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with too many items', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      assert.throws(() => {
-        schema.convertRow('string', '10.0', '1', 'string', 'string', 'string')
-      }, Error)
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with wrong type (fail fast)', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      assert.throws(() => {
-        schema.convertRow('string', 'notdecimal', '10.6', 'string', 'string',
-                          { failFast: true })
-      }, Error)
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert row with wrong type multiple errors', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      try {
-        schema.convertRow('string', 'notdecimal', '10.6', 'string',
-                          true)
-      } catch (e) {
-        assert.isArray(e)
-        assert.equal(e.length, 3)
-      }
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('convert multiple rows', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      const rows = schema.convert(
-        [
-          ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-          , ['string', '10.0', '1', 'string', 'string']
-        ])
-      for (const row of rows) {
-        assert.deepEqual(['string', '10.0', 1, 'string', 'string'],
-                         row)
-      }
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with wrong types fail fast', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      assert.throws(() => {
-        schema.convert(
-          [
-            ['string', 'not', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', 'an', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', '10.0', 'integer', 'string', 'string']
-          ], true)
-      }, Error)
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with wrong types multiple errors', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      try {
-        schema.convert([['string', 'not', '1', 'string', 'string']
-                         , ['string', '10.0', '1', 'string', 'string']
-                         , ['string', 'an', '1', 'string', 'string']
-                         , ['string', '10.0', '1', 'string', 'string']
-                         , ['string', '10.0', 'integer', 'string', 'string']])
-      } catch (e) {
-        assert.isArray(e)
-        assert.equal(e.length, 3)
-      }
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with incorrect row length fail fast', done => {
-    (new Schema(SCHEMA)).then(schema => {
-      assert.throws(() => {
-        schema.convert(
-          [
-            ['string', '10.0', '1', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string', 1]
-            , ['string', '10.0', '1', 'string', 'string']
-            , ['string', '10.0', '1', 'string', 'string']
-          ], true)
-      }, Error)
-      done()
-    }, (error) => {
-      assert.isNull(error)
-      done()
-    })
-  })
-
-  it('shouldn\'t convert rows with incorrect row length multiple errors',
-     done => {
-       (new Schema(SCHEMA)).then(schema => {
-         try {
-           schema.convert(
-             [['string', '10.0', '1', 'string']
-               , ['string', '10.0', '1', 'string', 'string']
-               , ['string', '10.0', '1', 'string', 'string', 1]
-               , ['string', '10.0', '1', 'string', 'string']
-               , ['string', '10.0', '1', 'string', 'string']])
-         } catch (e) {
-           assert.isArray(e)
-           assert.equal(e.length, 2)
-         }
-         done()
-       }, (error) => {
-         assert.isNull(error)
-         done()
-       })
-     })
-
   it('should load json file', done => {
     const url = 'http://localhost/remote.json'
     fetchMock.restore()
